Update main.c
So its kinda working now...still some bugs with the remove word function, and when you try to repeat playing the game, it exits. I think thats a scanf error but I couldn't figure out how to change that. I have more time on Saturday to work more on this!